Add files via upload
This project is based on the renting of the available cars at the shop. In this we have 2 domains one of the user and another is of the admin.
the user selects the car that he/she wants and make sure it is available. then he proceeds to the payment gateway where he needs to fill the card details to make a payment and is request will be sent to the admin.
then admin accepts or rejects the user request and car will be assigned to the user for the amount of period he has made a payment.
in this we have use html css js php and sqldatabases